Move observable APIs to a separate file.
